Split USB code off into device hander rs file
